requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-talisman==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
qrcode==7.4.2
# Prefer a Pillow release with prebuilt wheels for Windows/Python 3.13
Pillow==10.4.0
openpyxl==3.1.2
bcrypt==4.1.2
validators==0.22.0
supabase
